requirements.txt
aioredis==1.3.1
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.2.1
Automat==20.2.0
beautifulsoup4==4.9.3
Brotli==1.0.9
certifi==2021.5.30
cffi==1.14.6
channels==3.0.4
channels-redis==3.3.0
click==8.0.1
colorama==0.4.4
constantly==15.1.0
cryptography==3.4.7
cycler==0.11.0
daphne==3.0.2
dash==1.21.0
dash-bootstrap-components==0.13.0
dash-core-components==1.17.1
dash-html-components==1.1.4
dash-renderer==1.9.1
dash-table==4.12.0
dataclasses==0.8
Django==3.2.6
django-bootstrap4==3.0.1
django-plotly-dash==1.6.4
django-redis==5.0.0
dpd-components==0.1.0
dpd-static-support==0.0.5
Flask==2.0.1
Flask-Compress==1.10.1
future==0.18.2
hiredis==2.0.0
hyperlink==21.0.0
idna==3.2
importlib-metadata==2.1.1
incremental==21.3.0
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.3.4
msgpack==1.0.2
mysql-connector-python==8.0.29
mysqlclient==2.0.3
numpy==1.19.5
pandas==1.1.2
Pillow==8.4.0
plotly==5.2.2
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyMySQL==1.0.2
pyOpenSSL==20.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2021.1
redis==3.5.3
service-identity==21.1.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tenacity==8.0.1
Twisted==21.7.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing-extensions==3.10.0.0
Werkzeug==2.0.1
wincertstore==0.2
zipp==3.5.0
zope.interface==5.4.0
